+++ AntiDebug/DebuggerDetections.hpp
@@ -4,32 +4,34 @@
 class DebuggerDetections final  : public Debugger::AntiDebug
 {
 public:
-    DebuggerDetections(Settings* s, EvidenceLocker* evidence, shared_ptr<NetClient> netClient) : Debugger::AntiDebug(s, evidence, netClient)
+    DebuggerDetections(Settings* s, EvidenceLocker* evidence) : Debugger::AntiDebug(s, evidence)
     {
-        AddDetectionFunction([this]() -> bool { return _IsDebuggerPresent(); });
-        AddDetectionFunction([this]() -> bool { return _IsDebuggerPresent_HeapFlags(); });
-        AddDetectionFunction([this]() -> bool { return _IsDebuggerPresent_CloseHandle(); });
-        AddDetectionFunction([this]() -> bool { return _IsDebuggerPresent_RemoteDebugger(); });
-        AddDetectionFunction([this]() -> bool { return _IsDebuggerPresent_VEH(); });
-        AddDetectionFunction([this]() -> bool { return _IsDebuggerPresent_DbgBreak(); });
-        AddDetectionFunction([this]() -> bool { return _IsDebuggerPresent_PEB(); });
-        AddDetectionFunction([this]() -> bool { return _IsDebuggerPresent_DebugPort(); });
-        AddDetectionFunction([this]() -> bool { return _IsDebuggerPresent_ProcessDebugFlags(); });
-        AddDetectionFunction([this]() -> bool { return _IsKernelDebuggerPresent(); });
-        AddDetectionFunction([this]() -> bool { return _IsKernelDebuggerPresent_SharedKData(); });
-        //AddDetectionFunction([this]() -> bool { return _ExitCommonDebuggers(); });
+#ifndef _DEBUG //VS debugger messes up if threads get hidden from it
+        this->HideAllThreadsFromDebugger();
+#endif
+
+        AddDetectionFunction([this]() -> DetectionFlags { return _IsDebuggerPresent(); });
+        AddDetectionFunction([this]() -> DetectionFlags { return _IsDebuggerPresent_HeapFlags(); });
+        AddDetectionFunction([this]() -> DetectionFlags { return _IsDebuggerPresent_CloseHandle(); });
+        AddDetectionFunction([this]() -> DetectionFlags { return _IsDebuggerPresent_RemoteDebugger(); });
+        AddDetectionFunction([this]() -> DetectionFlags { return _IsDebuggerPresent_VEH(); });
+        AddDetectionFunction([this]() -> DetectionFlags { return _IsDebuggerPresent_PEB(); });
+        AddDetectionFunction([this]() -> DetectionFlags { return _IsDebuggerPresent_DebugPort(); });
+        AddDetectionFunction([this]() -> DetectionFlags { return _IsDebuggerPresent_ProcessDebugFlags(); });
+        AddDetectionFunction([this]() -> DetectionFlags { return _IsKernelDebuggerPresent(); });
+        AddDetectionFunction([this]() -> DetectionFlags { return _IsKernelDebuggerPresent_SharedKData(); });
+        //AddDetectionFunction([this]() -> DetectionFlags { return _ExitCommonDebuggers(); });
     }
 
-    bool _IsDebuggerPresent() { return IsDebuggerPresent(); }
-    bool _IsDebuggerPresent_HeapFlags();
-    bool _IsDebuggerPresent_CloseHandle();
-    bool _IsDebuggerPresent_RemoteDebugger();
-    bool _IsDebuggerPresent_VEH();
-    bool _IsDebuggerPresent_DbgBreak();
-    bool _IsDebuggerPresent_PEB();
-    bool _IsDebuggerPresent_DebugPort();
-    bool _IsDebuggerPresent_ProcessDebugFlags();
-    bool _IsKernelDebuggerPresent();
-    bool _IsKernelDebuggerPresent_SharedKData();
-    bool _ExitCommonDebuggers(); //call ExitProcess in a remote thread on common debuggers
+    DetectionFlags _IsDebuggerPresent();
+    DetectionFlags _IsDebuggerPresent_HeapFlags();
+    DetectionFlags _IsDebuggerPresent_CloseHandle();
+    DetectionFlags _IsDebuggerPresent_RemoteDebugger();
+    DetectionFlags _IsDebuggerPresent_VEH();
+    DetectionFlags _IsDebuggerPresent_PEB();
+    DetectionFlags _IsDebuggerPresent_DebugPort();
+    DetectionFlags _IsDebuggerPresent_ProcessDebugFlags();
+    DetectionFlags _IsKernelDebuggerPresent();
+    DetectionFlags _IsKernelDebuggerPresent_SharedKData();
+    DetectionFlags _ExitCommonDebuggers(); //call ExitProcess in a remote thread on common debuggers
 };