+++ Core/Detections.hpp
@@ -69,7 +69,8 @@ public:
 	static bool DoesFunctionAppearHooked(__in const char* moduleName, __in const char* functionName); //checks for jumps or calls as the first byte on a function
 	
 	static bool DoesIATContainHooked(); //check IAT for hooks
-	
+	static std::list<ProcessData::ImportFunction> FetchHookedIATEntries();
+
 	static uint64_t FindWritableAddress(__in const std::string moduleName, __in const std::string sectionName); //check all pages in .text section of image module for writable pages (after remapping, our .text section should only have RX protected pages)
 	
 	static bool CheckOpenHandles(); //detect any open handles to our process, very useful since this will detect most external cheats