+++ Obscure/VirtualMachineException.hpp
@@ -3,7 +3,12 @@
 #include <unordered_map>
 #include <string>
 
-// VMException class
+/**
+ * @brief Exception class for the Virtual Machine
+ *
+ * This class inherits from std::exception and provides specific error messages for various VM exceptions.
+ * It is used to handle errors that occur during the execution of the virtual machine's bytecode.
+ */
 class VirtualMachineException : public std::exception
 {
 public:
@@ -20,9 +25,13 @@ public:
 
     VirtualMachineException(VMException exception) : type(exception)
     {
-
     }
 
+    /**
+     * @brief Overridden method from std::exception, required for inheritance
+     *
+	 * @return a const char* pointer to the exception message
+    */
     const char* what() const noexcept override
     {
         return exceptionMessages.at(type).c_str();